+++ src/main/java/net/enzo/plus/Static.java
@@ -2,7 +2,7 @@ package net.enzo.plus;
 
 public class Static {
     public static final String MODID = "plus";
-    public static final String VERSION = "1.2";
+    public static final String VERSION = "1.3";
     public static final String COMMON = "net.enzo.plus.common.proxy.CommonProxy";
     public static final String CLIENT = "net.enzo.plus.client.proxy.ClientProxy";
 }